+++ crates/sequencer/src/runner.rs
@@ -748,8 +748,8 @@ where
             .unseal();
 
         let base_fee = calculate_next_block_base_fee(
-            latest_header.gas_used as u128,
-            latest_header.gas_limit as u128,
+            latest_header.gas_used,
+            latest_header.gas_limit,
             latest_header
                 .base_fee_per_gas
                 .expect("Base fee always set in Citrea"),
